Update to run as command
Updating to run the script as a command. The collection ID is passed in as the first parameter. The output is now written to the screen and user should redirect the screen output to a file if needed. README.md update to reflect updates and example of how to run command.